requirements.txt
python-telegram-bot==21.6
aiohttp==3.10.5
python-dotenv==1.1.1
matplotlib==3.8.2
seaborn==0.13.0
